added luggage test package
